doc-tagger-function/requirements.txt
azure.functions~=1.0a
markdown2
bs4
requests